requirements.txt
openai
python-dotenv==1.0.0
fastapi==0.95.1
uvicorn==0.22.0